Correlograms, coefficients and lags

These are the functions that can be used to calculate correlograms and define correlation coefficients and lags

Squiggles.correlogram — Function

correlogram(A, B, τ, threads_per_block)
correlogram(A, τ, threads_per_block)

Calculate the correlogram values for two signal matrices A (nsamples x nsignals) and B (nsamples x msignals). Each signal in A is run across each signal in B with lags ∈ [-τ, τ]

Arguments

A::AbstractArray – nsamples x nsignals matrix

B::AbstractArray – nsamples x msignals matrix

τ::Integer – Lag length

threads_per_block – Number of threads per block

Outputs

correlograms::AbstractArray{T, 3} – The correlogram volume of size (nsignals x msignals x τ*2+1).

Notes

Also one signal matrix can be used as input. In such cases it is assumed A⋆A thus to avoid extra computation, only correlations in the lower triangle will be carried out and the output will be of size N-1 x N/2 (for N even) or N X (N-1)/2 (for N odd).

As of now only signals that have n samples as a power of 2 can be used. If this condition is not met an error is thrown

Squiggles.norm_correlogram — Function

norm_correlogram(A, B, τ, threads_per_block)
norm_correlogram(A, τ, threads_per_block)

Calculate the normalized correlogram values for two signal matrices A (nsamples x nsignals) and B (nsamples x msignals). Each signal in A is run across each signal in B with lags ∈ [-τ, τ]

Arguments

A::AbstractArray – nsamples x nsignals matrix

B::AbstractArray – nsamples x msignals matrix

τ::Integer – Lag length

threads_per_block – Number of threads per block

Outputs

correlograms::AbstractArray{T, 3} – The correlogram volume fo size (nsignals x msignals x τ*2+1).

Notes

Also one signal matrix can be used as input. In such cases it is assumed A⋆A thus to avoid extra computation, only correlations in the lower triangle will be carried out and the output will be of size N-1 x N/2 (for N even) or N X (N-1)/2 (for N odd).

As of now only signals that have n samples as a power of 2 can be used. If this condition is not met an error is thrown

Squiggles.simplelags — Function

simplelags(correlograms, τ)
simplelags(correlograms, τ, sampling_rate)

Find the correlation or anticorrelation coeffs and correspective lag. These values correspond to the absolute maximum correlation value (preserving the original sign).

Arguments

correlograms::AbstractArray{T, 3} – Correlogram volume

τ::Integer – lag length

sampling_rate::Integer – Sampling rate

Outputs

coeffs::AbstractMatrix – Correlation coefficient matrix

lags::AbstractMatrix – Lag matrix

Notes

If the sampling rate is provided then the lags will be in seconds

Examples

simplelags(correlograms, 128) – Lags will be integers

simplelags(correlograms, 128, 100) – Lags will be in seconds

Squiggles.correlogram! — Function

correlogram!(correlograms, A, B, τ, threads_per_block)
correlogram!(correlograms, A, τ, threads_per_block)

Inplace version of the correlogram function.

Notes

This assumes that the inputs are already moved to the device, use prepare_inputs

Squiggles.norm_correlogram! — Function

norm_correlogram!(correlograms, A, B, τ, threads_per_block)
norm_correlogram!(correlograms, A, τ, threads_per_block)

Inplace version of the norm_correlogram function.

Notes

This assumes that the inputs are already moved to the device, use prepare_inputs

Squiggles.simplelags! — Function

simplelags!(coeffs, lags, correlograms, τ)
simplelags!(coeffs, lags, correlograms, τ, sampling_rate)

Inplace version of the simplelags function.

Notes

This assumes that the inputs are already moved to the device, use prepare_inputs