Available kernels

As of now, there are a fixed amount of kernels that can be used for specific signal lengths (power of 2). This page is mainly to show which signal lengths are supported, do not use these function to run correlation, use correlogram or norm_correlogram.

Squiggles.correlogram_ak16 — Function

correlogram_ak16(templates, signals, τ, correlograms)
correlogram_ak16(templates, τ, correlograms)

Calculate correlograms for templates and signals that have 16 samples

Arguments

templates::AbstractArray – Array of templates. This must be an array of 16 x ntemplates (rows x cols)

signals::AbstractArray – Array of signals. This must be an array of 16 x nsignals (rows x cols)

τ::Int – Length of samples to be correlated [-τ; τ].

correlograms::AbstractArray – Output array of correlograms. This must be of size ntemplates x nsignals x τ*2+1

Output

This functions works inplace and will overwrite data in the correlograms matrix

Notes

τ controls the total number of lags that will be performed for the correlation. The lags will be symmetrical in respect to τ = 0 and thus will be τ*2+1. For example, using τ = 16 will slide the template from τ = -16 to τ = 16.

Squiggles.correlogram_ak32 — Function

correlogram_ak32(templates, signals, τ, correlograms)
correlogram_ak32(templates, τ, correlograms)

Calculate correlograms for templates and signals that have 32 samples

Arguments

templates::AbstractArray – Array of templates. This must be an array of 32 x ntemplates (rows x cols)

signals::AbstractArray – Array of signals. This must be an array of 32 x nsignals (rows x cols)

τ::Int – Length of samples to be correlated [-τ; τ].

correlograms::AbstractArray – Output array of correlograms. This must be of size ntemplates x nsignals x τ*2+1

Output

This functions works inplace and will overwrite data in the correlograms matrix

Notes

τ controls the total number of lags that will be performed for the correlation. The lags will be symmetrical in respect to τ = 0 and thus will be τ*2+1. For example, using τ = 32 will slide the template from τ = -32 to τ = 32.

Squiggles.correlogram_ak64 — Function

correlogram_ak64(templates, signals, τ, correlograms)
correlogram_ak64(templates, τ, correlograms)

Calculate correlograms for templates and signals that have 64 samples

Arguments

templates::AbstractArray – Array of templates. This must be an array of 64 x ntemplates (rows x cols)

signals::AbstractArray – Array of signals. This must be an array of 64 x nsignals (rows x cols)

τ::Int – Length of samples to be correlated [-τ; τ].

correlograms::AbstractArray – Output array of correlograms. This must be of size ntemplates x nsignals x τ*2+1

Output

This functions works inplace and will overwrite data in the correlograms matrix

Notes

τ controls the total number of lags that will be performed for the correlation. The lags will be symmetrical in respect to τ = 0 and thus will be τ*2+1. For example, using τ = 64 will slide the template from τ = -64 to τ = 64.

Squiggles.correlogram_ak128 — Function

correlogram_ak128(templates, signals, τ, correlograms)
correlogram_ak128(templates, τ, correlograms)

Calculate correlograms for templates and signals that have 128 samples

Arguments

templates::AbstractArray – Array of templates. This must be an array of 128 x ntemplates (rows x cols)

signals::AbstractArray – Array of signals. This must be an array of 128 x nsignals (rows x cols)

τ::Int – Length of samples to be correlated [-τ; τ].

correlograms::AbstractArray – Output array of correlograms. This must be of size ntemplates x nsignals x τ*2+1

Output

This functions works inplace and will overwrite data in the correlograms matrix

Notes

τ controls the total number of lags that will be performed for the correlation. The lags will be symmetrical in respect to τ = 0 and thus will be τ*2+1. For example, using τ = 128 will slide the template from τ = -128 to τ = 128.

Squiggles.correlogram_ak512 — Function

correlogram_ak512(templates, signals, τ, correlograms)
correlogram_ak512(templates, τ, correlograms)

Calculate correlograms for templates and signals that have 512 samples

Arguments

templates::AbstractArray – Array of templates. This must be an array of 512 x ntemplates (rows x cols)

signals::AbstractArray – Array of signals. This must be an array of 512 x nsignals (rows x cols)

τ::Int – Length of samples to be correlated [-τ; τ].

correlograms::AbstractArray – Output array of correlograms. This must be of size ntemplates x nsignals x τ*2+1

Output

This functions works inplace and will overwrite data in the correlograms matrix

Notes

τ controls the total number of lags that will be performed for the correlation. The lags will be symmetrical in respect to τ = 0 and thus will be τ*2+1. For example, using τ = 512 will slide the template from τ = -512 to τ = 512.

Squiggles.correlogram_ak1024 — Function

correlogram_ak1024(templates, signals, τ, correlograms)
correlogram_ak1024(templates, τ, correlograms)

Calculate correlograms for templates and signals that have 1024 samples

Arguments

templates::AbstractArray – Array of templates. This must be an array of 1024 x ntemplates (rows x cols)

signals::AbstractArray – Array of signals. This must be an array of 1024 x nsignals (rows x cols)

τ::Int – Length of samples to be correlated [-τ; τ].

correlograms::AbstractArray – Output array of correlograms. This must be of size ntemplates x nsignals x τ*2+1

Output

This functions works inplace and will overwrite data in the correlograms matrix

Notes

τ controls the total number of lags that will be performed for the correlation. The lags are symmetrical in respect to τ = 0 and thus will be τ*2+1. For example, using τ = 1024 will slide the template from τ = -1024 to τ = 1024.