Utilities

Here are some functions that are used within the code but can be used also independently

Squiggles.memcopy — Function

memcopy(A::AbstractGPUArray)
memcopy(A::AbstractArray, dev=0)

Copy an array to and from GPU/CPU depending on the output data type.

Overload this method with GPU specific Arrays using extensions

Examples

using Squiggles
using CUDA

use_GPUbackend()

A = rand(256, 256)
memcopy(A) # Copy A to the default device of an available backend
memcopy(A, 1) # Copy A to device 1 of the available backend

Squiggles.prepare_inputs — Function

prepare_inputs(A, B, τ; device=0)
prepare_inputs(A, τ; device=0)

Preallocate the inputs for correlogram!, norm_correlogram! and simplelags! on device.

Arguments

A::AbstractArray{T} – Signal matrix A

B::AbstractArray{T} – Signal matrix B

τ::Integer – Lag length

device::Integer – Device to copy the data

Outputs

A_gpu::AbstractGPUArray{T} – Signal matrix A pointer on the device

B_gpu::AbstractGPUArray{T} – Signal matrix B pointer on the device

c_GPU::AbstractGPUArray{T, 3} – Empty correlogram volume pointer on the device

coeffs_GPU::AbstractGPUArray{T} – Empty correlation coefficient matrix pointer on the device

lags_GPU::AbstractGPUArray{T} – Empty lag matrix pointer on the device

Squiggles.get_available_GPUplatforms — Function

get_available_GPUplatforms()

Get the list of supported GPU platforms that can be used for running the simulations

Example

using Squiggles

get_available_GPUplatforms()

Squiggles.set_CPUbackend — Function

set_CPUbackend()

Set the backend to CPU

Example

using Squiggles
using CUDA

use_CPUbackend()

Squiggles.set_GPUbackend — Function

set_GPUbackend()

Set the backend to the available GPU.

Notes

The set backend is decided by the installed packages, i.e. if CUDA.jl is installed (and imported) then CUDA will be used. Only import one JuliaGPU package per script. If no JuliaGPU package is loaded, this method will return a LoadError. See the available GPU platforms by running getavailableGPUplatforms(). Some GPU backends support UnifiedMemory (e.g. CUDA, ROCm, Metal). The default will always be DeviceMemory but users can choose UnifiedMemory by running set_GPUbackend("unified")

Example

using Squiggles
using CUDA

use_GPUbackend()

Squiggles.linear2tr_nodiag — Function

linear2tr_nodiag(k, n)

Get the i, j index of the lower triangle (excluding the diagonal) of a matrix of size n x n from a linear index k. Matrix is col-wise ordered

Arguments

k::Integer – The linear index

n::Integer – The size of the matrix (n x n)

Notes

https://atrebas.github.io/post/2021-01-17-indextolowertriangularsubscripts/

Squiggles.reduce_M_nodiag — Function

reduce_M_nodiag(M)

Given a symmetric matrix M of size N x N (with N > 2), get a reduced matrix m of the lower triangle, excluding the diagonal. m will be of size N-1 x N/2 (for N even) or N X (N-1)/2 (for N odd).

Arguments

M::Matrix – Input matrix of NxN

Output

m::Matrix – Reduced matrix of N-1 x N/2 (for N even) or N X (N-1)/2 (for N odd).

Squiggles.reconstruct_nodiag — Function

reconstruct_nodiag(M)

Given a reduced matrix m of size N-1 x N/2 (for N even) or N X (N-1)/2 (for N odd), get a symmetrical matrix M of size N x N.

Arguments

m::Matrix – Reduced matrix of N-1 x N/2 (for N even) or N X (N-1)/2 (for N odd).

Output

M::Matrix – Input matrix of NxN

Squiggles.normalize_columns — Function

normalize_columns(M)

Normalize the columns of a matrix M (x/norm(x) where x is the column)

Squiggles.calculate_memory — Function

calculate_memory(A, B, τ)
calculate_memory(A, τ)

Calculate the memory required to perform A⋆B or A⋆A